requirements.txt
requests
duckdb
python-dateutil
pytest